// src/app/api/models/avatar-dto.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface AvatarDto {

  /**
   * ID del archivo en la tabla file_upload
   */
  id: number;

  /**
   * Tipo MIME del archivo
   */
  mimeType: string;

  /**
   * Nombre original del archivo
   */
  originalName: string;

  /**
   * Path relativo del archivo
   */
  path: string;

  /**
   * Tamaño del archivo en bytes
   */
  size: number;

  /**
   * URL del thumbnail (si existe)
   */
  thumbnailUrl?: string;

  /**
   * URL directa del archivo para mostrar en el frontend
   */
  url: string;
}